Extract depth-panel state capture/restore into engine/src/depth/panelState.ts
Moves captureDepthPanelStates, captureDepthFormValues,
restoreDepthPanelStates, and restoreDepthFormValues out of
PointCloudVisualizer - the logic that preserves each depth settings
panel's open/closed state and form values (fx/fy/cx/cy, distortion
coefficients, live-update flag) across file-list re-renders. 8 call
sites for capture/restore kept working unchanged via thin delegating
wrappers.

main.ts: 12,050 -> 11,851 lines. Now under 12k for the first time this
session (started at 15,008).

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
